benchmark foreachPar_ directly instead of foreachParN_
